Clarify documentation on Operator diff_burnable_mats
In PR #1392, it was revealed that the distribution of material
volumes with ``diff_burnable_mats`` was not clear. In that PR,
the Operator now divides the volume of the shared material
equally across all identical and newly created instances.
This commit adds some clarifying language into the Operator
docstring and depletion users guide on the handling
of volumes across repeated materials.