Allow Operator to reduce depletion chain from geometry
A new parameter, reduce_chain, can be used to instruct the
operator to create a smaller depletion chain. The geometry is
used to find all nuclides in burnable materials. The argument
can be a boolean, or a non-negative integer. A value of True
indicates no depth for following the paths, while an integer
can be used to directly specify the depth.