ResultList.get_atoms supports atom densities, days
Two new optional arguments to ResultsList.get_atoms are
introduced that control the units on the return values.
Time can be returned in second or in days if time_units is
"s" or "d", respectively. Concentrations can be returned
in atoms/cm^3 or atoms/b/cm if nuc_units is one of those values.
The default is to return time in seconds and number of atoms,
retaining back compatibility.