Configure depletion solver with Integrator argument, attribute
Users can pass a string or callable function to the Integrator
construction to configure the depletion solver, defaulting to
CRAM48. Supported strings are "cram16" and "cram48". Callables are
passed to the new solver attribute, and must accept three input
arguments for A, n0, and dt. Functions are expected to return
an array of compositions for this material.

Imports of CRAM48 and CRAM16 are delayed until inside the class
since the cram module imports from abc and can lead to a circular
import.