Made Current class consistent with mesh domain type
1) Kept by_nuclide, num_nuclides, and nuclides in the docstrings because when building a mgxs library for all mgxs_types, the variables should be the same accross all classes.
However, I noted that they were unused for a SurfaceMGXS.

2) Removed all code related to nuclides and micro xs since they were unused anyway.

3) Removed the possibility of computing a Current with any domain other than mesh.
Included an error message if a current mgxs object is instantiated with any domain other than mesh.
Since mesh domains and by_nuclide features are uncompatible, an existing message is already raised when mesh and by_nuclide are used together.

4) Changed estimator type to analog.