Finalized check for mesh & made progress on regression test
I had to put domain properties back into the SurfaceMGXS class to perform a checktype on the domain_type.
I updated the regression test mgxs_library_mesh to expect the correct input.
However, now the test fails for a 2-dimensional mesh. In 3D, the Current MGXS fails because it does not contain a 'z' dimension.
Meanwhile, all the other classes contain a 'z' dimension, equal to 1.
I need to figure out why the mesh used by SurfaceMGXS doesn't contain the same information as the other classes, even though it is the same mesh used in all.