Ordered the pandas dataframe
In the get_pandas_dataframe function of SurfaceMGXS, I removed all by_nuclide and micro options.
I also ordered the dataframe so that results would be ordered reliably.
This was necessary for tests to pass. Since the domain object is a little different for current than for the other classes,
I had to treat the reorder uniquely for the SurfaceMGXS class because of the presence of the mesh surface filter.
While other classes assume all meshes have a x, y, and z component, regardless of the dimension, this is not true in SurfaceMGXS. Therefore, I added an if statement for each dimension.