Change declaration order of maps used for vectors
If the destructor of an object tries to remove an entry from the map, there's no
guarantee it exists if the map has already been destroyed (because it was
declared after the vector of objects), resulting in a segfault. By declaring the
map first, we avoid this. Currently an issue for Nuclide, but potentially other
classes in the future.