Update build_xml to use serialization
File is largely based on the existing custom_source example.
Uses the serialization attribute on the source to
point to an XML file
containing the serialized representation of the source.
Also builds the serialized_source.xml file with some default values.
Uses a new name for the library containing the serializable source.