Create compatability with lattice cells
Previously, create_mg_mode assigned a new material to all cells in the geometry.
However, in certain cases, cells are filled with a lattice instead of a material.
Therefore, checking if a cell if filled with a material determines whether it is
appropriate to assign it a new multi group material or not.