Clean up criticality search (pls read comment)
Did a bit more for this one:
a) Got rid of warning messages that came from setting the root universe's ID to 0.
b) Moved source building to above the settings definition (rather than in the middle of it.)  
c) Removed `bracketed_method='bisect'` as this is the default argument.

Perhaps there should be a default tolerance as well? And is there a good reason to default `print_iterations` to false?