Incorporate feedback on settings definition
Also got rid of warnings by removing manually set ID's for cells. The ID for zirconium, for example was set to 2 which conflicted with an already defined material with an automatically generated ID of 2. Same for some others.

There is a part of the notebook that talks about ID's so I left that one alone. After the central point about ID's is made there's little point in keeping them in for the cells (unless you like looking at pink warning messages).