Replaced cell fix with a UserWarning in create_mg_mode
Since lattice or universe cells should still be able to have mgxs generated,
I repealed the requirement for the cell to be filled by a material.
Instead, if a non-material filled cell is used in the library domain,
a UserWarning warns the user not to include a consistuent cell along with
that lattice/universe cell.