Criticality search: return the minimum critical value
I've had a few experiences using `search_for_keff` where one of the guesses is closer to 1 than the final guess. This PR adjusts zero_value to correspond to the keff that is: 
a) critical (i.e. above 1) 
b) closest to sub-critical.

An alternative approach would be to return the value closest to the target.