Updated distribcell
This test often fails due to the formating of the results.
Now, I checked manually, and all these results are identical,
except of course for removing the "group" variable on the penultimate
chunk of results which is decay-rate.

So hopefully this formatting I get on my machine passes the other systems as well.