Cleaned up condense_xs
removed the if-statement pertaining to condense_dif_coef since
we assume that anyone who calls condense_xs() on DiffusionCoefficient
wants it condensed.

Also added an "else" to the if-statement pertaining to rxn_rate_tally is None.
In the test mgxs_library_condense, rxn_rate_tally is never called, so
the test fails if I remove that if-statement altogether. Instead,
adding an alternative in case rxn_rate_tally was already calculated
seemed like the best solution.