Update CMake required version number
This change increases the minimum version of CMake required from 3.3 to 3.22 for our GPU offloading version. The major motivation for this is to enforce the usage of the compiler preset file (as in older versions of cmake it will simply ignore the preset command and result in an incorrect configuration and user confusion). Presets were first introduced in 3.19, but I found when testing with 3.20 that it did not like some of the fields in our current preset file. 3.22 and newer all work.