Remove specific binaryDir location from preset
When presets were first introduced to CMake in 3.19, each preset required the `binaryDir` variable to be set to a specific build location. This precluded the user from building in an arbitrary directory of their choice. In CMake 3.21 (when version 3 of the preset schema was introduced) they removed this requirement. To allow more build flexibility, and to simplify the preset file, this PR simply removes all the `binaryDir` lines. Once incorporated, I'll have to update our auto testing scripts, but I think it is worthwhile as a user already ran into an issue with this.