Fix R < r torus intersection bug
In the case when R < r the inner "imaginary" surfaces created an additional source of scattering close to the center of the torus. This error occurred due to the lack of a check in the solution of the torus equation. If we
describe the equation in more detail, we obtain the following: . We see that the left part should obviously be less than or equal to zero. This check was missing in the code of surface.cpp file and was added in lines 972-978.