Description of Survival Normalization
Described Survival Normalization for phase-space sources. Unfortunately there is little description or justification for this out in the general topic literature. Though I have found it to work better in multiple test cases, notably the FOM in areas with middling statistics (~10%).

In term should add capability of normalization for biased general sources.