Add Fusion Optimized Multi-Group energy structure (FOMG-16k) Energy structures to calculate reaction rates with the multi-group binning (MGB) method.
https://fispact.ukaea.uk/wp-content/uploads/2016/07/snamc2013_05203.pdf

https://etheses.whiterose.ac.uk/id/eprint/4427/1/PhD_Thesis_Corrected_FINAL_FINAL_3_9_13.pdf

To-Do: add VESTA-43k energy structure

I believe I have followed the description of the structure as best I could:
"The optimal binning structure, named FOMG ( Fusion Optimised Multi Group) is split into three distinct sections. The thermal energy region is defined in the range 10−5eV ≤ EL < 1 eV with 1,000 bins and an energy width of 1.0 meV. The mid energy group is defined in the range 1 eV ≤ EL < 2 MeV with14,000 bins and a lethargy width of 1.036 × 10−3. The high energy group is defined in the range 2 MeV ≤ EH < 19.6 MeV with 1,000 bins and an energy width of 17.6 keV."