depletion: fix performance of chain mat construction
The previous construction was repeatedly invoking _validate_indices for
the scipy sparse matrix which does tons of expensive checks for every
square bracket indexing operation.  This was expensive and was causing
my full core depletion runs to take 10 hours with 90% of the time spent
in these validate indices checks.  This skips that and does a
batch/all-at-once construction of the sparse matrix after accumulating
all the indices and values.