make keff scaling consistent with keff normalization for kinetic simulations
Kinetic simulations based on an eigenvalue solve need to multiply the
prompt and delayed fission sources by the computed keff to bake-in the
steady state initial condition. The keff scaling introduced in PR 3595
complicates this by multiplying tallied flux by keff. This is fine,
however this new scaling factor is also applied to the final quantities
which serve as the final estimate of quantities for each timestep in
kinetic simulation. This would effetively undo the keff normalization
putting the system out of steady state. In practice this looks like a
shap step change in the values of tallied quantities. To address this,
this commit introduces control flow to ensure that the keff scaling
is applied to tallies at each timestep, but does not apply to the
internal quantitities stored in memory until the final timestep, to
ensure the keff normalization still bakes in that steady state.